Move test files to inside src
