+++ src/index.ts
@@ -269,4 +269,5 @@ export class OAuth2Strategy<
 	}
 }
 
-export { OAuth2RequestError, TokenResponseBody } from "oslo/oauth2";
+export { OAuth2RequestError } from "oslo/oauth2";
+export type { TokenResponseBody } from "oslo/oauth2";
