+++ tsconfig.json
@@ -12,5 +12,5 @@
 		"outDir": "./build"
 	},
 	"exclude": ["node_modules"],
-	"include": ["src/**/*.ts", "src/**/*.tsx"]
+	"include": ["src/index.ts"]
 }
